Rename module realization_state to realization_storage_state
