+++ tests/unit_tests/gui/tools/test_case_tool.py
@@ -7,7 +7,7 @@ from ert.gui.ertnotifier import ErtNotifier
 from ert.gui.tools.manage_cases.case_init_configuration import (
     CaseInitializationConfigurationPanel,
 )
-from ert.realization_state import RealizationStorageState
+from ert.realization_storage_state import RealizationStorageState
 
 
 @pytest.mark.usefixtures("copy_poly_case")
