+++ tests/unit_tests/simulator/test_simulation_context.py
@@ -1,5 +1,5 @@
 from ert.enkf_main import EnKFMain
-from ert.realization_state import RealizationStorageState
+from ert.realization_storage_state import RealizationStorageState
 from ert.simulator import SimulationContext
 from tests.utils import wait_until
