+++ tests/unit_tests/status/test_tracking_integration.py
@@ -30,7 +30,7 @@ from ert.ensemble_evaluator.state import (
     JOB_STATE_START,
     REALIZATION_STATE_FINISHED,
 )
-from ert.realization_state import RealizationStorageState
+from ert.realization_storage_state import RealizationStorageState
 from ert.shared.feature_toggling import FeatureToggling
